+++ src/utils/mult_modulo.c
@@ -1,0 +1,16 @@
+#include <netinet/in.h>
+#include "mult_modulo.h"
+
+uint64_t MultModulo(uint64_t a, uint64_t b, uint64_t mod) {
+  uint64_t result = 0;
+  a = a % mod;
+  while (b > 0) {
+    if (b % 2 == 1)
+      result = (result + a) % mod;
+    a = (a * 2) % mod;
+    b /= 2;
+  }
+
+  return result % mod;
+}
+
