+++ src/utils/mult_modulo.h
@@ -1,0 +1,6 @@
+#ifndef MULT_MODULO_H_
+#define MULT_MODULO_H_
+
+uint64_t MultModulo( uint64_t a, uint64_t b, uint64_t mod);
+
+#endif //   MULT_MODULO_H_
